WIP: Change config from JSON to YAML
This enables me to put comments in the config file. I could skip the
middleman and expect the user to edit Makefile.variables, but (a) that
would be far too simple and (b) this makes the config a little more
portable for when I inevitably rewrite the Makefile in another language.

Here's a neat trick for converting an existing JSON file to YAML:

    jq . config.json | yq --yaml-output > config.yaml